+++ benchmarks/ExtremeScaling/bunny.cxx
@@ -24,7 +24,7 @@
 #include <t8.h>
 #include <t8_cmesh.h>
 #include <t8_cmesh_tetgen.h>
-#include <t8_cmesh_vtk.h>
+#include <t8_cmesh_vtk_writer.h>
 #include <t8_cmesh/t8_cmesh_examples.h>
 #include <t8_forest.h>
 #include <t8_schemes/t8_default/t8_default_cxx.hxx>
