+++ benchmarks/time_forest_partition.cxx
@@ -28,7 +28,7 @@
 #include <p4est_connectivity.h>
 #include <t8_cmesh.h>
 #include <t8_cmesh/t8_cmesh_examples.h>
-#include <t8_cmesh_vtk.h>
+#include <t8_cmesh_vtk_writer.h>
 #include <t8_cmesh/t8_cmesh_partition.h>
 #include <t8_cmesh/t8_cmesh_occ.hxx>
 #include <t8_cmesh_readmshfile.h>
