+++ example/IO/cmesh/gmsh/t8_load_and_refine_square_w_hole.cxx
@@ -27,7 +27,7 @@
 
 #include <sc_refcount.h>
 #include <t8_cmesh.h>
-#include <t8_cmesh_vtk.h>
+#include <t8_cmesh_vtk_writer.h>
 #include <t8_cmesh/t8_cmesh_partition.h>
 #include <t8_cmesh_readmshfile.h>
 #include <t8_forest.h>
