+++ example/IO/cmesh/gmsh/t8_read_msh_file.cxx
@@ -22,10 +22,9 @@
 
 #include <sc_options.h>
 #include <t8.h>
-#include <t8_cmesh.h>
+#include <t8_cmesh_vtk_writer.h>
 #include <t8_cmesh/t8_cmesh_partition.h>
 #include <t8_cmesh_readmshfile.h>
-#include <t8_cmesh_vtk.h>
 #include <t8_schemes/t8_default/t8_default_cxx.hxx>
 
 /* Output a cmesh in .vtk format. Process i writes to the file
