+++ example/IO/cmesh/tetgen/t8_forest_from_tetgen.cxx
@@ -25,7 +25,7 @@
 #include <t8_cmesh.h>
 #include <t8_cmesh/t8_cmesh_partition.h>
 #include <t8_cmesh_tetgen.h>
-#include <t8_cmesh_vtk.h>
+#include <t8_cmesh_vtk_writer.h>
 #include <t8_schemes/t8_default/t8_default_cxx.hxx>
 #include <t8_forest.h>
