+++ example/forest/t8_test_face_iterate.cxx
@@ -29,7 +29,7 @@
 #include <t8_forest.h>
 #include <t8_cmesh.h>
 #include <t8_cmesh_readmshfile.h>
-#include <t8_cmesh_vtk.h>
+#include <t8_cmesh_vtk_writer.h>
 #include <t8_cmesh/t8_cmesh_examples.h>
 #include <t8_data/t8_containers.h>
