+++ example/forest/t8_test_ghost.cxx
@@ -28,7 +28,7 @@
 #include <t8_forest.h>
 #include <t8_cmesh.h>
 #include <t8_cmesh_readmshfile.h>
-#include <t8_cmesh_vtk.h>
+#include <t8_cmesh_vtk_writer.h>
 #include <t8_cmesh/t8_cmesh_examples.h>
 #include <example/common/t8_example_common.h>
