+++ example/forest/t8_test_ghost_large_level_diff.cxx
@@ -37,7 +37,7 @@
 #include <sc_statistics.h>
 #include <sc_options.h>
 #include <t8_cmesh_readmshfile.h>
-#include <t8_cmesh_vtk.h>
+#include <t8_cmesh_vtk_writer.h>
 #include <t8_cmesh/t8_cmesh_examples.h>
 #include <t8_schemes/t8_default/t8_default_prism/t8_dprism.h>
 #include <t8_schemes/t8_default/t8_default_tet/t8_dtet.h>
