+++ test/t8_geometry/t8_test_point_inside.cxx
@@ -24,7 +24,7 @@
 #include <sc_functions.h>
 #include <t8_eclass.h>
 #include <t8_cmesh.h>
-#include <t8_cmesh_vtk.h>
+#include <t8_cmesh_vtk_writer.h>
 #include <t8_cmesh/t8_cmesh_examples.h>
 #include <t8_forest.h>
 #include <t8_forest/t8_forest_iterate.h>
